session: 세션에 관리자 유무 속성 추가
isMaster 미들웨어 주석처리함/ login_id는 패킷으로 요청하지 않도록 처리

+++ src/api/routes/user.js
@@ -36,31 +36,47 @@ module.exports = (app) => {
   });
 
   // 회원정보 수정
-  // TODO : 세션 관리 필요
+  // 세션 관리 완
   router.post("/update", async (req, res, next) => {
     try {
-      const login_id = req.body.login_id;
-      const userInput = req.body.modification;
+      if (req.session.logined) {
+        const login_id = req.session.login_id;
+        const userInput = req.body.modification;
+        //console.log(login_id);
+  
+        const user = await user_service.updateUser(login_id, userInput);
+  
+        return res.status(200).json({ success: true, response: user });
+      } else {
+        console.log("[-] /update :: 로그인 세션이 존재하지 않습니다.");
+        return res.status(201).json({ success: false });
+      }
 
-      const user = await user_service.updateUser(login_id, userInput);
-
-      return res.status(200).json({ success: true, response: user });
     } catch (e) {
       res.status(400).json({ success: false, errorMsg: e.message });
     }
   });
 
   // 회원 탈퇴
-  // TODO : 세션 관리 필요
+  // 세션 관리 완
+  //router.post("/delete", async (req, res, next) => {
   router.delete("/", async (req, res, next) => {
     try {
-      const login_id = req.body.login_id;
+      if (req.session.logined) {
+        const login_id = req.session.login_id;
+        //console.log(login_id);
 
-      const isDeleted = await user_service.deleteUser(login_id);
+        const isDeleted = await user_service.deleteUser(login_id);
+  
+        return res
+          .status(isDeleted ? 200 : 400)
+          .json({ success: isDeleted, response: { isDeleted: isDeleted } });
+      } else {
+        console.log("[-] delete :: 로그인 세션이 존재하지 않습니다.");
+        return res.status(201).json({ success: false });
+      }
 
-      return res
-        .status(isDeleted ? 200 : 400)
-        .json({ success: isDeleted, response: { isDeleted: isDeleted } });
+
     } catch (e) {
       res.status(400).json({ success: false, errorMsg: e.message });
     }
@@ -68,30 +84,43 @@ module.exports = (app) => {
 
   //===================================================================================
   // 관리자_회원강제 탈퇴
-  router.delete("/admin", isMaster, async (req, res, next) => {
+  router.post("/admin/delete", async (req, res, next) => {
+  //router.delete("/admin", isMaster, async (req, res, next) => {
     try {
-      const target_userId = req.body.user_login_id;
+      if (req.session.admin && req.session.logined) {
+        const target_userId = req.body.user_login_id;
 
-      const isDeleted = await user_service.deleteUser_admin(target_userId);
+        const isDeleted = await user_service.deleteUser_admin(target_userId);
+  
+        return res
+          .status(isDeleted ? 200 : 400)
+          .json({ success: isDeleted, response: { isDeleted: isDeleted } });
+      } else {
+        console.log("[-] 관리자_회원강제 탈퇴 :: 로그인 세션이 존재하지 않습니다.");
+        return res.status(201).json({ success: false });
+      }
 
-      return res
-        .status(isDeleted ? 200 : 400)
-        .json({ success: isDeleted, response: { isDeleted: isDeleted } });
     } catch (e) {
       res.status(400).json({ success: false, errorMsg: e.message });
     }
   });
 
   // 관리자_회원탈퇴 복구
-  router.post("/admin/restoration", isMaster, async (req, res, next) => {
+  router.post("/admin/restoration", async (req, res, next) => {
     try {
-      const target_userId = req.body.user_login_id;
+      if (req.session.admin && req.session.logined) {
+        const target_userId = req.body.user_login_id;
+        console.log(target_userId);
 
-      const isRestored = await user_service.reStoreUser_admin(target_userId);
-
-      return res
-        .status(isRestored ? 200 : 400)
-        .json({ success: true, response: { isRestored: isRestored } });
+        const isRestored = await user_service.reStoreUser_admin(target_userId);
+  
+        return res
+          .status(isRestored ? 200 : 400)
+          .json({ success: true, response: { isRestored: isRestored } });
+      } else {
+        console.log("[-] 관리자_회원강제 복구 :: 로그인 세션이 존재하지 않습니다.");
+        return res.status(201).json({ success: false });
+      }
     } catch (e) {
       res.status(400).json({ success: false, errorMsg: e.message });
     }
